Add parsing for template variables
